+++ src/contexts/APIContext/Provider.tsx
@@ -1,11 +1,13 @@
 import React, { useCallback } from "react";
 import Context from "./Context";
+import { IGetAccountResult, IUnsignedTransaction } from "types/NXTAPI";
 import getAccount from "utils/api/getAccount";
 import sendJUP from "utils/api/sendJUP";
 import getAccountId from "utils/api/getAccountId";
 import getBlockchainStatus from "utils/api/getBlockchainStatus";
 import getBalance from "utils/api/getBalance";
-import { IGetAccountResult, IUnsignedTransaction } from "types/NXTAPI";
+import getBlockchainTransactions from "utils/api/getBlockchainTransactions";
+import setAccountInfo from "utils/api/setAccountInfo";
 
 const APIProvider: React.FC = ({ children }) => {
   const handleGetAccount = useCallback(async (address: string) => {
@@ -24,14 +26,28 @@ const APIProvider: React.FC = ({ children }) => {
     return sendJUP(unsignedTxJSON); // return the result back to the caller so they can work with the whole signed object/error for now
   }, []);
 
+  const handleGetBlockchainTransactions = useCallback(async (account: string) => {
+    let transactions;
+    try {
+      transactions = await getBlockchainTransactions(account);
+    } catch (e) {
+      console.error("error getting transactions in API provider");
+      return false;
+    }
+
+    return transactions;
+  }, []);
+
   return (
     <Context.Provider
       value={{
         getBlockchainStatus,
         getAccount: handleGetAccount,
+        setAccountInfo,
         getAccountId,
         getBalance,
         sendJUP: handleSendJUP,
+        getMyTxs: handleGetBlockchainTransactions,
       }}
     >
       {children}