+++ src/hooks/useMyTxs.ts
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import { MyTxContext } from "../contexts/MyTxContext";
+
+const useMyTxs = () => ({
+  ...useContext(MyTxContext),
+});
+
+export default useMyTxs;