+++ src/views/Dashboard/components/Widgets/SendWidget/SendWidget.tsx
@@ -1,18 +1,18 @@
 import React, { memo, useCallback, useMemo, useState } from "react";
-import { Autocomplete, Box, Button, Dialog, FormGroup, Grid, Input, styled, Typography } from "@mui/material";
+import { Autocomplete, Box, Button, DialogContent, FormGroup, Grid, Input, Stack, styled, Typography } from "@mui/material";
 import TextField from "@mui/material/TextField";
 import { IUnsignedTransaction } from "types/NXTAPI";
 import { isValidAddress } from "utils/validation";
 import useAccount from "hooks/useAccount";
 import useAPI from "hooks/useAPI";
+import { JUPGenesisTimestamp, standardDeadline, standardFee } from "utils/common/constants";
+import JUPDialog from "components/JUPDialog";
+
+// [x]: Pagination
+// MUST: Improve styling
+// [x]: Shortcut to page
 
 const placeHolderVals = ["JUP", "ASTRO"];
-
-const JUPGenesisTimestamp = 1508627969; // can be found in getConstants() API call as "epochBeginning"
-
-// TODO: implement as advanced features?
-const standardFee = "5000";
-const standardDeadline = 1440;
 
 const SendWidget: React.FC = () => {
   const [toAddress, setToAddress] = useState<string>("");
@@ -21,6 +21,10 @@ const SendWidget: React.FC = () => {
   const [userSecretInput, setUserSecretInput] = useState<string>("");
   const { accountRs, publicKey } = useAccount();
   const { sendJUP, getAccount, getAccountId } = useAPI();
+
+  const handleCloseSeedCollection = useCallback(() => {
+    setRequestUserSecret(false);
+  }, []);
 
   const fetchRecipAccountId = useCallback(async () => {
     if (getAccount !== undefined && getAccountId !== undefined) {
@@ -112,83 +116,89 @@ const SendWidget: React.FC = () => {
   const ConditionalSendWidget = useMemo(() => {
     return requestUserSecret ? (
       <>
-        <Dialog open={requestUserSecret}>
-          <Box sx={{ minWidth: "600px", height: "300px" }}>
-            <Typography align="center">Please enter your seed phrase.</Typography>
-            <Input onChange={(e) => handleSecretEntry(e.target.value)} type="password" placeholder="Enter Seed Phrase"></Input>
-            <Button variant="contained" onClick={() => handleSubmitSecret(userSecretInput)}>
-              Confirm & Send
-            </Button>
-            <Button variant="outlined" onClick={() => setRequestUserSecret(false)}>
-              Cancel
-            </Button>
-          </Box>
-        </Dialog>
+        <JUPDialog isOpen={requestUserSecret} closeFn={handleCloseSeedCollection}>
+          <DialogContent>
+            <Box sx={{ minWidth: "600px", height: "300px" }}>
+              <Typography align="center">Please enter your seed phrase.</Typography>
+              <Stack sx={{ alignItems: "center" }}>
+                <SeedphraseEntryBox onChange={(e) => handleSecretEntry(e.target.value)} type="password" placeholder="Enter Seed Phrase" />
+                <ConfirmButton variant="contained" onClick={() => handleSubmitSecret(userSecretInput)}>
+                  Confirm & Send
+                </ConfirmButton>
+              </Stack>
+            </Box>
+          </DialogContent>
+        </JUPDialog>
       </>
     ) : (
-      <Box sx={{ border: "1px dotted green", margin: "10px", height: "300px" }}>
-        <FormGroup>
-          <Grid container>
-            <Grid container>
-              <Grid item xs={12}>
-                <StyledWidgetHeading>Send JUP</StyledWidgetHeading>
-              </Grid>
-              <Grid item xs={12}>
-                <StyledAutocomplete
-                  freeSolo
-                  options={placeHolderVals.map((option) => option)}
-                  renderInput={(params) => <TextField {...params} label="Enter asset name" />}
-                />
-              </Grid>
-              <Grid item xs={12}>
-                <StyledToAddressInput onChange={(e) => handleToAddressEntry(e.target.value)} placeholder="To Address" />
-              </Grid>
-              <Grid item xs={12}>
-                <StyledQuantityInput onChange={(e) => handleQuantityEntry(e.target.value)} placeholder="Quantity" />
-              </Grid>
-            </Grid>
-            <Grid container>
-              <Grid item xs={12}>
-                <StyledSendButton fullWidth onClick={handleSend} variant="contained">
-                  Send
-                </StyledSendButton>
-              </Grid>
-            </Grid>
+      <>
+        <StyledWidgetHeading>Send JUP</StyledWidgetHeading>
+
+        <Grid container>
+          <Grid item xs={10}>
+            <StyledAutocomplete
+              freeSolo
+              options={placeHolderVals.map((option) => option)}
+              renderInput={(params) => <TextField {...params} label="Enter asset name" />}
+            />
+            <StyledToAddressInput onChange={(e) => handleToAddressEntry(e.target.value)} placeholder="To Address" />
+            <StyledQuantityInput onChange={(e) => handleQuantityEntry(e.target.value)} placeholder="Quantity" />
           </Grid>
-        </FormGroup>
-      </Box>
+          <Grid item xs={2}>
+            <StyledSendButton fullWidth onClick={handleSend} variant="contained">
+              Send
+            </StyledSendButton>
+          </Grid>
+        </Grid>
+      </>
     );
-  }, [handleQuantityEntry, handleSecretEntry, handleSend, handleSubmitSecret, handleToAddressEntry, requestUserSecret, userSecretInput]);
+  }, [
+    handleCloseSeedCollection,
+    handleQuantityEntry,
+    handleSecretEntry,
+    handleSend,
+    handleSubmitSecret,
+    handleToAddressEntry,
+    requestUserSecret,
+    userSecretInput,
+  ]);
 
   return <>{ConditionalSendWidget}</>;
 };
+
+const SeedphraseEntryBox = styled(Input)(({ theme }) => ({
+  minWidth: "400px",
+  margin: "40px 0px",
+}));
+
+const ConfirmButton = styled(Button)(({ theme }) => ({
+  margin: "20px 0px",
+}));
 
 const StyledWidgetHeading = styled(Typography)(() => ({
   textAlign: "center",
 }));
 
 const StyledAutocomplete = styled(Autocomplete)(() => ({
-  minWidth: "250px",
+  width: "90%",
   padding: "10px",
+  margin: "0px 10px",
 }));
 
 const StyledToAddressInput = styled(Input)(() => ({
-  minWidth: "550px",
+  width: "90%",
   padding: "10px",
-  margin: "10px",
+  margin: "10px 10px",
 }));
 
-// TODO: find out how to fill width, still not 100% decided on this component's base
 const StyledQuantityInput = styled(Input)(() => ({
-  minWidth: "550px",
+  width: "90%",
   padding: "10px",
-  margin: "10px",
+  margin: "10px 10px",
 }));
 
-// TODO: find out how to get the height to auto fill
 const StyledSendButton = styled(Button)(() => ({
-  margin: "10px",
-  minHeight: "250px",
+  height: "100%",
 }));
 
 export default memo(SendWidget);